[refactor] chat 컨트롤러 패키지 묶음, 사소한 컨트롤러, 엔티티 리팩터링
